pyproject.toml
[project]
name = "rag-functionalbase"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.15",
    "langchain-text-splitters>=1.1.1",
    "numpy>=2.4.2",
    "pymupdf>=1.27.1",
    "pypdf>=6.7.2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.3",
]
